Rename Form to BaseForm and StatefulForm to Form for clarity
