+++ packages/form-for/src/index.js
@@ -1,9 +1,9 @@
 // @flow
 
-import Form from './Form';
-export type { Props as FormProps, SchemaProperty, Schema } from './Form';
+import BaseForm from './BaseForm';
+export type { Props as FormProps, SchemaProperty, Schema } from './BaseForm';
 
-import StatefulForm from './StatefulForm';
+import Form from './Form';
 
 import FieldGroup from './FieldGroup';
 export type { Props as FieldGroupProps } from './FieldGroup';
@@ -13,4 +13,4 @@ export type { ComponentProps, Props as FieldProps } from './Field';
 
 import fieldDecorator from './fieldDecorator';
 
-export { Form, StatefulForm, Field, fieldDecorator as field, FieldGroup };
+export { BaseForm, Form, Field, fieldDecorator as field, FieldGroup };
